+++ go.mod
@@ -168,3 +168,5 @@ require (
 	modernc.org/memory v1.5.0 // indirect
 	modernc.org/sqlite v1.23.1 // indirect
 )
+
+replace github.com/goravel/framework => ../goravel/framework
