+++ view_test.go
@@ -1,10 +1,380 @@
 package fiber
 
 import (
+	"io"
+	"net/http"
 	"testing"
 
+	configmocks "github.com/goravel/framework/contracts/config/mocks"
+	contractshttp "github.com/goravel/framework/contracts/http"
+	httpmocks "github.com/goravel/framework/contracts/http/mocks"
+	"github.com/goravel/framework/support/file"
 	"github.com/stretchr/testify/assert"
 )
+
+func TestView_Make(t *testing.T) {
+	var (
+		err        error
+		fiber      *Route
+		req        *http.Request
+		mockConfig *configmocks.Config
+		mockView   *httpmocks.View
+	)
+
+	assert.Nil(t, file.Create("resources/views/empty.tmpl", `{{ define "empty.tmpl" }}
+1
+{{ end }}
+`))
+	assert.Nil(t, file.Create("resources/views/data.tmpl", `{{ define "data.tmpl" }}
+{{ .Name }}
+{{ .Age }}
+{{ end }}
+`))
+
+	beforeEach := func() {
+		mockConfig = &configmocks.Config{}
+		mockConfig.On("GetBool", "http.drivers.fiber.prefork", false).Return(false).Once()
+		ConfigFacade = mockConfig
+
+		mockView = &httpmocks.View{}
+		ViewFacade = mockView
+	}
+	tests := []struct {
+		name        string
+		method      string
+		url         string
+		setup       func(method, url string) error
+		expectCode  int
+		expectBody  string
+		expectPanic bool
+	}{
+		{
+			name:   "data is empty, shared is empty",
+			method: "GET",
+			url:    "/make/empty",
+			setup: func(method, url string) error {
+				mockView.On("GetShared").Return(nil).Once()
+
+				fiber.Get("/make/empty", func(ctx contractshttp.Context) {
+					ctx.Response().View().Make("empty.tmpl")
+				})
+
+				var err error
+				req, err = http.NewRequest(method, url, nil)
+				if err != nil {
+					return err
+				}
+
+				return nil
+			},
+			expectCode: http.StatusOK,
+			expectBody: "\n1\n",
+		},
+		{
+			name:   "data is empty, shared is not empty",
+			method: "GET",
+			url:    "/make/data",
+			setup: func(method, url string) error {
+				mockView.On("GetShared").Return(map[string]any{
+					"Name": "test",
+					"Age":  18,
+				}).Once()
+
+				fiber.Get("/make/data", func(ctx contractshttp.Context) {
+					ctx.Response().View().Make("data.tmpl")
+				})
+
+				var err error
+				req, err = http.NewRequest(method, url, nil)
+				if err != nil {
+					return err
+				}
+
+				return nil
+			},
+			expectCode: http.StatusOK,
+			expectBody: "\ntest\n18\n",
+		},
+		{
+			name:   "data is not empty, shared is not empty",
+			method: "GET",
+			url:    "/make/data",
+			setup: func(method, url string) error {
+				mockView.On("GetShared").Return(map[string]any{
+					"Name": "test",
+				}).Once()
+
+				fiber.Get("/make/data", func(ctx contractshttp.Context) {
+					ctx.Response().View().Make("data.tmpl", map[string]any{
+						"Age": 18,
+					})
+				})
+
+				var err error
+				req, err = http.NewRequest(method, url, nil)
+				if err != nil {
+					return err
+				}
+
+				return nil
+			},
+			expectCode: http.StatusOK,
+			expectBody: "\ntest\n18\n",
+		},
+		{
+			name:   "data is not empty, shared is not empty, and data contains shared key",
+			method: "GET",
+			url:    "/make/data",
+			setup: func(method, url string) error {
+				mockView.On("GetShared").Return(map[string]any{
+					"Name": "test",
+				}).Once()
+
+				fiber.Get("/make/data", func(ctx contractshttp.Context) {
+					ctx.Response().View().Make("data.tmpl", map[string]any{
+						"Name": "test1",
+						"Age":  18,
+					})
+				})
+
+				var err error
+				req, err = http.NewRequest(method, url, nil)
+				if err != nil {
+					return err
+				}
+
+				return nil
+			},
+			expectCode: http.StatusOK,
+			expectBody: "\ntest1\n18\n",
+		},
+		{
+			name:   "data is struct, shared is not empty, and data contains shared key",
+			method: "GET",
+			url:    "/make/data",
+			setup: func(method, url string) error {
+				mockView.On("GetShared").Return(map[string]any{
+					"Name": "test",
+				}).Once()
+
+				fiber.Get("/make/data", func(ctx contractshttp.Context) {
+					ctx.Response().View().Make("data.tmpl", struct {
+						Name string
+						Age  int
+					}{
+						Name: "test1",
+						Age:  18,
+					})
+				})
+
+				var err error
+				req, err = http.NewRequest(method, url, nil)
+				if err != nil {
+					return err
+				}
+
+				return nil
+			},
+			expectCode: http.StatusOK,
+			expectBody: "\ntest1\n18\n",
+		},
+		{
+			name:   "data is []string",
+			method: "GET",
+			url:    "/make/data",
+			setup: func(method, url string) error {
+				mockView.On("GetShared").Return(nil).Once()
+
+				fiber.Get("/make/data", func(ctx contractshttp.Context) {
+					assert.Panics(t, func() {
+						ctx.Response().View().Make("data.tmpl", []string{"test"})
+					})
+				})
+
+				var err error
+				req, err = http.NewRequest(method, url, nil)
+				if err != nil {
+					return err
+				}
+
+				return nil
+			},
+			expectCode: http.StatusOK,
+			expectBody: "",
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			beforeEach()
+			fiber, err = NewRoute(mockConfig, nil)
+			assert.Nil(t, err)
+
+			err := test.setup(test.method, test.url)
+			assert.Nil(t, err)
+
+			resp, err := fiber.Test(req)
+			assert.NoError(t, err)
+
+			if test.expectBody != "" {
+				body, err := io.ReadAll(resp.Body)
+				assert.Nil(t, err)
+				assert.Equal(t, test.expectBody, string(body))
+			}
+
+			assert.Equal(t, test.expectCode, resp.StatusCode)
+
+			mockConfig.AssertExpectations(t)
+			mockView.AssertExpectations(t)
+		})
+	}
+
+	assert.Nil(t, file.Remove("resources"))
+}
+
+func TestView_First(t *testing.T) {
+	var (
+		err        error
+		fiber      *Route
+		req        *http.Request
+		mockConfig *configmocks.Config
+		mockView   *httpmocks.View
+	)
+
+	assert.Nil(t, file.Create("resources/views/empty.tmpl", `{{ define "empty.tmpl" }}
+1
+{{ end }}
+`))
+	assert.Nil(t, file.Create("resources/views/data.tmpl", `{{ define "data.tmpl" }}
+{{ .Name }}
+{{ .Age }}
+{{ end }}
+`))
+
+	beforeEach := func() {
+		mockConfig = &configmocks.Config{}
+		mockConfig.On("GetBool", "http.drivers.fiber.prefork", false).Return(false).Once()
+		ConfigFacade = mockConfig
+
+		mockView = &httpmocks.View{}
+		ViewFacade = mockView
+	}
+	tests := []struct {
+		name        string
+		method      string
+		url         string
+		setup       func(method, url string) error
+		expectCode  int
+		expectBody  string
+		expectPanic bool
+	}{
+		{
+			name:   "found the first view",
+			method: "GET",
+			url:    "/first",
+			setup: func(method, url string) error {
+				mockView.On("Exists", "empty.tmpl").Return(true).Once()
+				mockView.On("GetShared").Return(nil).Once()
+
+				fiber.Get("/first", func(ctx contractshttp.Context) {
+					ctx.Response().View().First([]string{"empty.tmpl", "data.tmpl"})
+				})
+
+				var err error
+				req, err = http.NewRequest(method, url, nil)
+				if err != nil {
+					return err
+				}
+
+				return nil
+			},
+			expectCode: http.StatusOK,
+			expectBody: "\n1\n",
+		},
+		{
+			name:   "found the second view",
+			method: "GET",
+			url:    "/first",
+			setup: func(method, url string) error {
+				mockView.On("Exists", "empty.tmpl").Return(false).Once()
+				mockView.On("Exists", "data.tmpl").Return(true).Once()
+				mockView.On("GetShared").Return(nil).Once()
+
+				fiber.Get("/first", func(ctx contractshttp.Context) {
+					ctx.Response().View().First([]string{"empty.tmpl", "data.tmpl"}, map[string]any{
+						"Name": "test",
+						"Age":  18,
+					})
+				})
+
+				var err error
+				req, err = http.NewRequest(method, url, nil)
+				if err != nil {
+					return err
+				}
+
+				return nil
+			},
+			expectCode: http.StatusOK,
+			expectBody: "\ntest\n18\n",
+		},
+		{
+			name:   "no view found",
+			method: "GET",
+			url:    "/first",
+			setup: func(method, url string) error {
+				mockView.On("Exists", "empty.tmpl").Return(false).Once()
+				mockView.On("Exists", "data.tmpl").Return(false).Once()
+
+				fiber.Get("/first", func(ctx contractshttp.Context) {
+					assert.Panics(t, func() {
+						ctx.Response().View().First([]string{"empty.tmpl", "data.tmpl"}, map[string]any{
+							"Name": "test",
+							"Age":  18,
+						})
+					})
+				})
+
+				var err error
+				req, err = http.NewRequest(method, url, nil)
+				if err != nil {
+					return err
+				}
+
+				return nil
+			},
+			expectCode: http.StatusOK,
+			expectBody: "",
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			beforeEach()
+			fiber, err = NewRoute(mockConfig, nil)
+			assert.Nil(t, err)
+
+			err := test.setup(test.method, test.url)
+			assert.Nil(t, err)
+
+			resp, err := fiber.Test(req)
+			assert.NoError(t, err)
+
+			if test.expectBody != "" {
+				body, err := io.ReadAll(resp.Body)
+				assert.Nil(t, err)
+				assert.Equal(t, test.expectBody, string(body))
+			}
+
+			assert.Equal(t, test.expectCode, resp.StatusCode)
+
+			mockConfig.AssertExpectations(t)
+			mockView.AssertExpectations(t)
+		})
+	}
+
+	assert.Nil(t, file.Remove("resources"))
+}
 
 func TestStructToMap(t *testing.T) {
 	data := struct {
